python_agent/requirements.txt
crewai
groq
flask==3.0.2
python-dotenv==1.0.1
